Add support to array-like vec3
In current JSON format 3D vectors are also given as
strings separated by whitespace, with this commit they
can also be given as [number, number, number] format.

This was unnecessary for the time being but I've added
for the sake of practice.

Also this is duplicate code but the same is implemented for
f64 datatype.