Add compute_normal( ) to Triangle
This is the naive implementation I believe and
warning, it is not tested yet. There are some commented
code of my attempts to create structures for faster memory
access but this can be a premature optimization so I either
will delete it soon or implement after naive version.

Next goal is to learn how to write unit tests in Rust so that
I can test this function and future ones.