Add TriangleNormals, remove Triangle.normal
Dealing with geometric structures often come with a cost
of huge structs holding data that may or may not be useful
throughout the application. So in order to avoid bloated
structs, here I will stick with Libigl convention (which is
also adopted by gpytoolbox), that is to stick with simple
matrices. So here I declare TriangleNormals which is
supposed to be struct of arrays. Not me stepping out of
the decision at the first step... according to libigl convention
I would be simply writing a function like
triangle_normals(V, F) and return N that is Matrix-like.

Actually there are some modifications I want to do, like
TriangleNormals can actually be more generic because it
is simply holding coordinate-like data, which could be
used for vertices as well and I've created for loop as if
it is Python but Rust's mappings could be useful here.