Generalize TriangleNormals to SoA
Now things get Rusty, here a trait called
Vector3StructofArrays is declared (which I thought I'd
do it in later stages but couldn't help but experiment)
Oh wait, this is useless. Trait already assumes Vector3-like
data so it is no different than CoordLike struct I've defined.
It does work right now but it is not actually useful if I want to
implement another struct for this trait.