Add w, v, u vectors to Camera
+ Implement new( ) for Camera and Nearplane
(note that these are not meant to be used in source but they
were useful for tests in camera.rs and it is idiomatic Rust
to provide new( ) afaik.

One thing I'm not sure is declaration of these three vecs
inside Camera. Now my code needs to call .setup( ) to
fill these variables before it can use camera, might not be
a wise design choice.