Change image resolution data type to usize
I was using Int as the data type for image resolution but
negative resolution does not make sense so I've decided
to change it. However this came up with a cost of
generalizing deser_arr2 (whose name is acutally ambigious
that should be renamed) that can handle generic types
rather than depending on Int type.