Fix previous commit's error
So the simplest fix is apparently to bound generic T to
Clone. This same error trips me up so frequently though,
whenever I attempt to use generics I run into unbounded
trait issues. So the point is we don't know if given generic
type implements Clone or not, and we need to ensure it
does implement, otherwise it cannot call all( ) function
on that SingleOrVec enum (right?)