Convert Material trait to enum
I have used Material trait before to convert RT in One
Weekend book to Rust (and it was also used in RT Road to
Rust implementation) but I wonder what if we had enums
and define radiance with matching this enum. So in this
commit I briefly experiment with changing code structure
before I implement any ray tracing part.