Keep default values if a field is not found
+ Add set_optional( ) which sets scene member to json
value if the field is found, otherwise keep the default value

Some json files do not include all the expected fields.
My previous attempt constructed scene to default at first
but then if a field was not found it overrides the fields to
None but why would we want to change default to None?
e.g. it'd change background color to None (which is currently
None by default but in future it would be something different)