Remove scale from ImageData
I've just realized we don't need that scale parameter from ImageData,
I thought it'd be useful if width and height was given in centimeters etc.
but it is too early to think about images in a image editor sense. Here we
just care about number of pixels in width and height. (So I've also added a
note for future self)