Add new( ) and at( ) to Ray
Now I understand even better why providing new( ) is a way to not
turn every member of this struct public. I could've just made the interior
elements public so that I could call Ray { origin, direction } but this is not a
good practice for many cases. For example current Camera struct has no
useful methods to access certain elements, I've just declared them as
pub, probably not a good idea. I'm also not sure if I should return a Vector3
in at( ) -> Vector3, or if it should be boxed vector... probably since Vector3
is small, it is ok.