Create ray.rs, hitrecord
This module is intended to be responsible from Ray generation and its related
structs such as HitRecord which was presented in Ray Tracing in One Weekend
book. Here I have taken the struct I've used to implement this book earlier
this summer but considering if this struct could've been different. It was
useful to keep track of the last hit object info without cluttering other
structs.