Change return value of intersects_with
We want to return the data related to intersection or return None if
intersection did not happen. I guess I used something else in my
previous ray tracer attempt or it was this but the book implementation was
different? Not sure if it's better to create an empty hit record and pass it by
reference here or create a hit record in the function and return it.
Having verbalized it, I think it'd be better to pass a reference, because
in this way it will return an unnecessary copy. Could we return a pointer
to HitRecord? Then probably we need to specify the lifetime of this
HitRecord in rust.