Introduce ArrayOfTriangles (!)
This is the start of this experimental implementation. Since Intersectable
trait has function signature only taking ray and t, but our deserialized Triangle which is currently named TriangleSerde (awkward naming there) only holds indices referring to VertexData, it is not possible to directly implement Intersectable with TriangleSerde. By the way, Intersectable is
basically Hittable trait in RT in One Weekend book, I just thought Hittable
is confusing at least to me. Anyway so the solution I will attempt from now
on is to have a struct that can be used for both array of triangles literally,
and meshes. Hold on, this is no different than mesh struct, we could just
use Mesh struct and map array of triangles to that struct? Either way I think
that struct needs to hold V and F data so that we can implement functions
on these easier. But what about a single triangle?