Rename Triangle structs and introduce lifetime (!)
So I want to have a way to point related vertex data. One way would be
to introduce vertexData in intersection function arguments but that
breaks the expectation of traits, because I want to call intersection with the
same function signature for different shapes. Either all of the shapes will
pass vertexdata to that function (which doesn't sound that bad?) or
create a helper struct Triangle to hold the address of vertexdata. On the
downside, why every single triangle will hold the same pointer? it sounds
unnecessary. Lifetime marker is added to explicitly state that Triangle should
not outlive the vertexdata, so vertexdata is as valid as long as Triangle afaik.
Having verbalized, I think passing vertexdata in function signature could be
more neat. I was thinking that spheres have no vertex data but - oh wait.
Spheres need vertex data as well, their center is defined by the index of
vertexdata.