Insert dummy vertex to shift vertex data
So every index referring to VertexData is assumed to be starting from 1.
This is an issue if we like to directly use ids on vertexData, I thought we can
shift faces to make them start from 0, but then there are so many fields with
the same assumption, if one is forgotten to be shifted, then there will a bug
to search for. Then I thought we could just append an origin vertex in the
beginning to solve the problem. Yet this would be a bug in future if one
would like to load some other object and if the faces start from 0. So I've
changed the name of the function setup( ) to setup_after_json( ) to explicitly
remind myself this is only valid for our json loading logic.