Convert triangle indices field to fixed array (!)
Since triangle shape will always have exactly 3 vertices for indices, I wanted
to switch to a regular array instead of heap allocated vector.
Also introduced some variables in intersects_with but there is a problem,
we verts is not a vector slice, we cannot access coordinates with [ ] directly.
I could use _data [ ] but I wonder if we can override operators (?) in Rust
before I proceed on that.