Create a template for LightContext
So here is the problem: in my current implementation in order to get the
final radiance for the material, information about light sources are needed but
I don't want to couple Material with SceneLights. From the slides, it seems like
we need a common set of information to compute shading, that is w_o, w_i,
light intensity etc. but just feeding it to the function looks verbose.
As a solution could we introduce a struct to hold relevant information
and give it to the materials? Well, it still couples materials with lights but
now there is a mediator LightContext which can be built upon SceneLights
and in future could be constructed from different Structs if needed. Since
we don't know the concrete material type in compile time, it seems like we
need to provide that get_color( ) function to the material trait otherwise
get_colorİ( ) in renderer need to infer Material type to decide which shading
to implement, but that gets messier in Rust, you cannot just infer it directly.
Apparently there are ways to "downcast" dyn Material to concrete Material
type, like with "Any" trait but it adds unnecessary complexity. Anyway, so
I've decided to put light context in material's radiance( ) function (or could
be renaed to get_color( ) here we have radiance to define color)