Define LightContext::new_from( )
I think this is a reasonable template for material to compute radiance
from LightContext, and there are two methods to be implemented:
ambient_radiance( ) and radiance( )
I was planning to merge these two but then the for loops should change,
LightContext would hold a vector of shadow rays, in short it'd complicate
its intention. So for now LightContext belongs to a single point light.
Ambient light although it has "light" in its name, it has a different job I think
so I don't want to merge it with the shading computations inside materials