Fix shadow acne
The problem was not on intersection but on epsilon. When epsilon is not
present in the file, silent defaulting makes it 0. The solution applied here
is to set them in scene.setup_after_json( ) but see my comments for more
thoughts on it.

- Comment out simple ray-triangle intersecion (waiting to be fixed), use
möller's algorithm instead
- Add debug_assert for sanity
- Comment out debug message in clamping (it was getting in the way
of seeing loaded scene, in fact I could only see the problem when I run
in debug mode, the scene literally had 0.0 for epsilon)
- There is also a commented else there, I'll remove it later just keeping it
in the commit to remind myself what I've used to debug shadow rays