Refactor get_color for recursion (!)
This is a broken commit, I've just moved shadow ray parts for diffuse
shading out of the function. Also I was having trouble with inferring
concrete type of dyn Material in complier time (what am I, rustc?) anyway
a simple solution is to implement get_type( ) for all material types and then
use match expression to decide which color to add! Currently it is
empty but for mirror or other types, it'll call get_color( ) with depth param
for diffuse materials it will call shade_diffuse (I believe for path tracing
that'd be replaced with get_color too)