Add depth to get_color( )
Here begins recursion! And we already hit yet another wall, what color we
should return at the base case of this recursion? I believe we need that
any_hit( ) function call in shade_diffuse but I'm not sure how to compute the
received irradiance?  No wait we have to compute reflection direction, just
like we computed shadow rays for diffuse. Then I think shadow ray
generation could be the part of diffuse material, and reflected ray would be
part of Mirror material and so materials could have a function to tell the
outgoing ray(s) so that we can compute their shading?