Fix comment
It was misleading by claiming *it never panics*, it does panic at the useful
time! Though since I've spent unnecessary amount of time to decide
to panic or not to panic, it is probably not panic in future. In either case,
that was a nice practice on useful error messages (is it really useful though?)