Remove radiance from materials
So far I've been mixing up terms! In renderer, get_color( ) is responsible
from returning radiance along a ray (that's what we do in ray tracing after
all!) but then in materials, there was radiance( ) function previously it was
returning radiance, then I changed it to return only the coefficients to
multiply received irradiance, so it basically became "attenuation" in
RTIOW book (right?). So to make it clear I've renamed the functions and
adjust get_ambient( ) that was taking light intensity as a parameter,
which is now decoupled.