Remove attenuation from reflect( ) signatures
This is now cleaner but we still have a problem, mirror's attenuate( )
returns it reflectance coefficient, and that looks really ambigious. Yes
that coefficient scales radiance but attenuation for dielectric material
is dependent on rays (and with this commit I've added ray_in and t to
attenuate( ) signature, so I believe the reflectance coefficient should be
used somewhere else.