Add conductor material (*)
THIS A MILESTONE! Cornellbox recursive works! It was easier than I expected
simply implemented Fresnel reflection and the rest of the logic was the same
with dielectric (thanks for starting from dielectric first!) So now I gotta test
for bigger scenes, and there is still .ply parsing missing for some input also
I want to use Rayon for multithreading but that will require BoxedMaterial
to be Rc oh wait to be Arc actually, then it will probably break my code
because I've done silly coding assuming material will be boxed... I go
now hide in my box of shame.