Change function signatures to accept VertexCache
Actually HeapAllocatedVerts that is Arc<VertexCache>
I now understand what is *spaghetti* about spaghetti code! Every time
I need a new data structure to be used within ray tracer flow, I need to
pass the same data through multiple functions, just so that it reaches the
target destination. This is not good, and it will not get resolved unless
serious refactoring on my structs and traits.